MERGE ALL WEB PROJECTS
This was a hard merge of all web projects related to Emmersiv West Health Institute
work that I did from 06/14-06/15. I concatenated the projects related to AngularJS
template, emmersiv webpage, and highcharts for results visualization into this
single directory.

Updated it with a gitignore file and all READMEs to be better.